+++ include/algorithms/variable_candidates_heuristic_antidote_search.hpp
@@ -42,7 +42,7 @@ namespace ghost
 		public:
 			VariableCandidatesHeuristicAntidoteSearch();
 			
-			std::vector<double> compute_variable_candidates( const SearchUnitData& data ) const override;
+			std::vector<int> compute_variable_candidates( const SearchUnitData& data ) const override;
 		};
 	}
 }
