+++ include/algorithms/variable_heuristic_uniform.hpp
@@ -42,7 +42,7 @@ namespace ghost
 		public:
 			VariableHeuristicUniform();
 			
-			int select_variable( const std::vector<double>& candidates, const SearchUnitData& data, randutils::mt19937_rng& rng ) const override;
+			int select_variable( const std::vector<int>& candidates, const SearchUnitData& data, randutils::mt19937_rng& rng ) const override;
 		};
 	}
 }
