+++ src/algorithms/variable_heuristic_uniform.cpp
@@ -35,7 +35,7 @@ VariableHeuristicUniform::VariableHeuristicUniform()
 	: VariableHeuristic( "Uniform" )
 { }
 
-int VariableHeuristicUniform::select_variable( const std::vector<double>& candidates, const SearchUnitData& data, randutils::mt19937_rng& rng ) const
+int VariableHeuristicUniform::select_variable( const std::vector<int>& candidates, const SearchUnitData& data, randutils::mt19937_rng& rng ) const
 {
-	return static_cast<int>( rng.pick( candidates ) );
+	return rng.pick( candidates );
 }
